+++ docs/changes.md
@@ -1,4 +1,8 @@
 Changes of joker-aligner
 ========================
 
+#### 0.0.3
+
+* modules renamed
+* Alignment class